# Improve Indicator Performance Benchmark