### What does this implement/fix? Explain your changes.
- Adding new `interop` `type` cache holding a `bool`, true if its an `exception`.
- Adding a `setter` and `getter` cache for the `propertyobject`.

#### How Has This Been Tested?
Tested exeucting `QuantConnect` unit and regression tests and indicator performance benchmark:

[MASTER]
![imagen](https://user-images.githubusercontent.com/18473240/53108754-4f72ff80-3516-11e9-901b-de06264d822a.png)
[PR]
![imagen](https://user-images.githubusercontent.com/18473240/53109365-7c73e200-3517-11e9-8ba7-058858e05dd7.png)

[MASTER]
160.95 seconds at 5k data points per second. Processing total of 782,223 data points.
158.54 seconds at 5k data points per second. Processing total of 782,223 data points.
159.95 seconds at 5k data points per second. Processing total of 782,223 data points.
[PR]
132.27 seconds at 6k data points per second. Processing total of 782,223 data points.
130.07 seconds at 6k data points per second. Processing total of 782,223 data points.
129.88 seconds at 6k data points per second. Processing total of 782,223 data points.

[PR] has a 15-20% speed improvement.

**Note**
Testing showed no signficant impact on the `BasicTemplateAlgorithm` performance benchmark:
[python_benchmark_algo.txt](https://github.com/QuantConnect/pythonnet/files/2885668/python_benchmark_algo.txt)


### Does this close any currently open issues?
Related to https://github.com/QuantConnect/Lean/issues/2925